+++ packages/cli/CHANGELOG.md
@@ -5,6 +5,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Improve handling of Windows CRLF line-endings 
+
 ## [0.1.0] - 2024-04-22
 
 Initial release.